+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [Unreleased](https://github.com/Piturnah/gex/compare/v0.3.3...master)
 ### Added
+- Warning when opening Gex with locale other than English ([#13](https://github.com/Piturnah/gex/issues/13))
 - Colour coding of `--help` flag output
 ### Changed
 - Minibuffer now maintains a stack of messages so messages are not lost if more than one is sent per frame